refactor: isolate Tabs stories
